+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Adds `showWeekends` flag in month view to hide & show weekends view. 
   Default is `showWeekends = true` shows all weekdays. [#385](https://github.com/SimformSolutionsPvtLtd/flutter_calendar_view/issues/385)
 - Events are now hidden for days not in the current month when hideDaysNotInMonth = true
+- Fixes right icon always shows default icon in `CalendarPageHeader` when providing custom icon. [#432](https://github.com/SimformSolutionsPvtLtd/flutter_calendar_view/issues/432)
 
 # [1.3.0 - 12 Nov 2024](https://github.com/SimformSolutionsPvtLtd/flutter_calendar_view/tree/1.3.0)